Add backup and restore scripts to cron service

+++ bin/cron.sh
@@ -1,6 +1,14 @@
 #!/bin/sh
 set -e
 
-cp /tmp/www-data.crontab /var/spool/cron/crontabs/www-data
-chown root:root /var/spool/cron/crontabs/www-data
+crontab=/tmp/www-data.crontab
+
+if  [ -f "${crontab}" ]
+then
+  cp /tmp/www-data.crontab /var/spool/cron/crontabs/www-data
+  chown root:root /var/spool/cron/crontabs/www-data
+fi
+
+apk add mariadb-client
+
 crond -f -d 8
